+++ Units/parser-kotlin.r/kotlin-multilinestrings.d/input.kt
@@ -1,5 +1,0 @@
-var used1 = """var unused1 = 1"""
-var used2 = """
-    var unused2 = 2
-"""
-var used3 = """var unused3 = 3""" + """var unused4 = 4"""