+++ Units/parser-kotlin.r/kotlin-rawstrings.d/input.kt
@@ -1,0 +1,9 @@
+var used1 = """var unused1 = 1"""
+var used2 = """
+    var unused2 = 2
+"""
+var used3 = """var unused3a = 3""" + """var unused3b = 3"""
+var used4 = """..."val unused4"..."""
+var used5 = """${process("""var unused5 = 5""")}"""
+var used6 = """"val unused6=6""""
+var used7 = """""var unused6=6"""""